rearrange: rename store folder to redux
