+++ src/index.tsx
@@ -13,7 +13,7 @@ import { Provider } from 'react-redux';
 
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import { persistor, store } from './store/store';
+import { persistor, store } from './redux/store';
 import { PersistGate } from 'redux-persist/integration/react';
 
 const theme = createTheme({
